refactor(domain): package 구조 수정 (line package 추가 및 도메인 객체 이동)
