+++ src/main/java/nextstep/subway/domain/line/Line.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.line;
 
 import java.util.Objects;
 import javax.persistence.Column;
@@ -7,6 +7,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import nextstep.subway.domain.BaseEntity;
 import nextstep.subway.dto.UpdateLine;
 
 @Entity
