+++ src/test/java/nextstep/subway/line/LineAcceptanceTest.java
@@ -11,7 +11,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import nextstep.subway.TestUtil;
-import nextstep.subway.domain.Line;
+import nextstep.subway.domain.line.Line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
